Adds the Angular router to the webapp, and factors out the test job code into its own component.
