+++ api-service/webapp-demo/src/app/app.component.html
@@ -1,85 +1,3 @@
-<!--The content below is only a placeholder and can be replaced.-->
-<h1>Conversation AI Crowdsourcing Demo</h1>
-
-<div class="counts">
-  <div>my worker id: {{userNonce}}</div>
-  <div>local_sent_count: {{local_sent_count}}</div>
-  <div>training_answer_count: {{training_answer_count}}</div>
-  <div>user_mean_score: {{user_mean_score}}</div>
-  <div>overall_job_answer_count: {{overall_job_answer_count}}</div>
-  <div>overall_job_mean_score: {{overall_job_mean_score}}</div>
-</div>
-
-<div *ngIf="errorMessage" class="errorMessage" (click)="clearError()">{{errorMessage}}</div>
-
-<div *ngIf="!question">
-  Loading...
-</div>
-
-<div *ngIf="question">
-  <div class="id">{{question.id}}</div>
-  <div class="text-header">Comment to rate:</div>
-  <div class="text">{{question.text}}</div>
-
-  <mat-checkbox [(ngModel)]="readableAndInEnglish">The comment is in English and understandable</mat-checkbox>
-
-  <div *ngIf="readableAndInEnglish" class="questions">
-    <div class="three-question-group">
-      <mat-radio-group [(ngModel)]="toxicityAnswer">
-        <div class="toxicity-selected-value"><span class="label-heading">Toxic?</span></div>
-        <div><mat-radio-button class="toxicity-radio-button" value="very">Very</mat-radio-button></div>
-        <div><mat-radio-button class="toxicity-radio-button" value="somewhat">Somewhat</mat-radio-button></div>
-        <div><mat-radio-button class="toxicity-radio-button" value="notatall">Not at all</mat-radio-button></div>
-      </mat-radio-group>
-    </div>
-
-    <div class="three-question-group">
-      <mat-radio-group [(ngModel)]="insultAnswer">
-        <div class="insult-selected-value"><span class="label-heading">Insulting?</span></div>
-        <div><mat-radio-button class="insult-radio-button" value="very">Very</mat-radio-button></div>
-        <div><mat-radio-button class="insult-radio-button" value="somewhat">Somewhat</mat-radio-button></div>
-        <div><mat-radio-button class="insult-radio-button" value="notatall">Not at all</mat-radio-button></div>
-      </mat-radio-group>
-    </div>
-
-    <div class="three-question-group">
-      <mat-radio-group [(ngModel)]="obsceneAnswer">
-        <div class="obscene-selected-value"><span class="label-heading">Obscene?</span></div>
-        <div><mat-radio-button class="obscene-radio-button" value="very">Very</mat-radio-button></div>
-        <div><mat-radio-button class="obscene-radio-button" value="somewhat">Somewhat</mat-radio-button></div>
-        <div><mat-radio-button class="obscene-radio-button" value="notatall">Not at all</mat-radio-button></div>
-      </mat-radio-group>
-    </div>
-
-    <div class="three-question-group">
-      <mat-radio-group [(ngModel)]="threatAnswer">
-        <div class="threat-selected-value"><span class="label-heading">Threatening?</span></div>
-        <div><mat-radio-button class="threat-radio-button" value="very">Very</mat-radio-button></div>
-        <div><mat-radio-button class="threat-radio-button" value="somewhat">Somewhat</mat-radio-button></div>
-        <div><mat-radio-button class="threat-radio-button" value="notatall">Not at all</mat-radio-button></div>
-      </mat-radio-group>
-    </div>
-
-    <div class="three-question-group">
-      <mat-radio-group [(ngModel)]="hateAnswer">
-        <div class="hate-selected-value"><span class="label-heading">Hateful to an identity?</span></div>
-        <div><mat-radio-button class="hate-radio-button" value="very">Very</mat-radio-button></div>
-        <div><mat-radio-button class="hate-radio-button" value="somewhat">Somewhat</mat-radio-button></div>
-        <div><mat-radio-button class="hate-radio-button" value="notatall">Not at all</mat-radio-button></div>
-      </mat-radio-group>
-    </div>
-  </div>
-
-  <div class="remarks">
-    <form class="example-form">
-    <mat-form-field class="example-full-width">
-        <textarea matInput placeholder="Other comments and reactions"></textarea>
-    </mat-form-field>
-    </form>
-  </div>
-
-  <div>
-    <button mat-raised-button color="primary" (click)="sendScoreToApi()">Submit</button>
-  </div>
-
+<div>
+  <router-outlet></router-outlet>
 </div>
