+++ api-service/webapp-demo/src/app/test_job.component.css
@@ -1,0 +1,54 @@
+
+.text-header {
+  font-weight: bold;
+}
+
+.text {
+  padding: 10px;
+  border: 1px solid #CCC;
+  margin: 10px;
+}
+
+.counts {
+  text-align: right;
+  padding: 5px;
+  font-size: 80%;
+}
+
+.id {
+  text-align: right;
+  padding: 5px;
+  font-size: 80%;
+}
+
+.questions {
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+}
+
+.three-question-group {
+  margin: 10px;
+}
+
+.label-heading{
+  font-weight: bold;
+}
+
+.example-form {
+  min-width: 150px;
+  max-width: 500px;
+  width: 100%;
+}
+
+.example-full-width {
+  width: 100%;
+}
+
+.errorMessage {
+  font-weight: bold;
+  padding: 10px;
+  border: 1px solid #FCC;
+  margin: 10px;
+  color: #933;
+}
